Fix selected conditions not showing up issue, fix infinite loops for create and update timestamps and vehicle dropdown options
